PubMedTools/CF/requirements.txt
openpyxl==3.0.3
lxml==4.4.2
crossrefapi==1.4.0
urllib3==1.25.3
more-itertools==7.2.0
spacy==2.1.8
biopython==1.74
pandas==0.25.0
python-dotenv==0.10.3
bert-serving-client==1.9.6
bert-serving-server==1.9.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.17.0
tensorflow-estimator==1.14.0
requests-oauthlib==1.2.0
selenium==3.141.0
requests==2.20.0
tensorflow-gpu==1.14.0
nltk==3.4.5
beautifulsoup4==4.8.1
bert-serving-client==1.9.8
bert-serving-server==1.9.8
scipy==1.3.1
